+++ src/ttsim/interface_dag_elements/labels.py
@@ -102,6 +102,7 @@ def top_level_namespace(
 @input_dependent_interface_function(
     include_if_any_input_present=[
         "input_data__flat",
+        "input_data__qname",
         "input_data__tree",
         "input_data__df_with_nested_columns",
         "input_data__df_and_mapper__df",
@@ -117,6 +118,7 @@ def input_columns_from_input_data(processed_data: QNameData) -> UnorderedQNames:
 @input_dependent_interface_function(
     include_if_no_input_present=[
         "input_data__flat",
+        "input_data__qname",
         "input_data__tree",
         "input_data__df_with_nested_columns",
         "input_data__df_and_mapper__df",